Refactoring...
