+++ kafka-start/.idea/compiler.xml
@@ -6,7 +6,8 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="kafka-start-course" />
+        <module name="kafka-producer-twitter" />
+        <module name="kafka-basics" />
       </profile>
     </annotationProcessing>
   </component>
