+++ kafka-start/kafka-basics/src/main/java/kafka/tutorial/ProducerDemo.java
@@ -1,4 +1,4 @@
-package com.github.jinnycho503.kafka.tutorial;
+package kafka.tutorial;
 
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerConfig;
