+++ kafka-start/kafka-producer-twitter/kafka-producer-twitter.iml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4" />
