+++ kafka-start/kafka-producer-twitter/src/main/java/kafka/twitter/ElasticSearchConsumer.java
@@ -1,11 +1,10 @@
-package com.github.jinnycho503.kafka.twitter;
+package kafka.twitter;
 
 import org.apache.http.HttpHost;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.elasticsearch.client.RestClient;
-import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
