+++ kafka-start/kafka-producer-twitter/src/main/java/kafka/twitter/TwitterProducer.java
@@ -1,4 +1,4 @@
-package com.github.jinnycho503.kafka.twitter;
+package kafka.twitter;
 
 import com.google.common.collect.Lists;
 import com.twitter.hbc.ClientBuilder;
